Pearson Dictionary API
This Android studio program uses the Dictionary API to get definitions
for user inputted words. The user is only provided with definitions that
exactly match the word for which they have searched.(i.e. When user
searches for "box," the API would return the definition for "boxed. This
program uses the API in such a fashion that only if there is an exact
match will a definition be returned.  The original output from an HTTPS
call for "box" can be seen in the "Result of BOX API Call.txt" file in
the root level of the Dictionary Demo). Also, the user is able to get
alternate definitions if available.